requirements.txt
# pip install python-telegram-bot --pre
scipy
numpy
scikit-learn
Pillow
telegram
matplotlib